+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 All notable changes to the project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [development version]
+## [gtk-fortran-extra 1.1.0] 2021-11-13
 
 ### Added
+- `unknown_pleasures/`: a new example to generate figures similar to the iconic Unknown Pleasures cover.
 - `my_fast_app/` and `my_long_app/`: margins at left and/or right of the TextView widgets, in order they do not touch the window's borders.
 - `my_long_app/`: the TextView is now automatically scrolling.